+++ breeze-spring-boot/src/main/kotlin/com/windea/breezeframework/springboot/data/DataAutoConfiguration.kt
@@ -1,6 +1,7 @@
 package com.windea.breezeframework.springboot.data
 
 import com.windea.breezeframework.springboot.data.components.converters.*
+import com.windea.breezeframework.springboot.data.converters.*
 import org.springframework.boot.autoconfigure.condition.*
 import org.springframework.context.annotation.*
 import org.springframework.data.domain.*